requirements.txt
jsonpath==0.82
pytest==5.1.1
selenium==3.141.0
requests==2.25.1
allure-pytest==2.8.31
configparser==5.0.1
jsonschema==2.6.0

PyJWT==1.7.1

pytest-cov==2.11.1
pytest-dependency==0.5.1
pytest-ordering==0.6
filelock==3.0.12
pyyaml==5.4.1
python-keycloak==0.24.0
